+++ R/highlight.R
@@ -75,6 +75,11 @@ highlight <- function(text, classes = classes_chroma(), pre_class = NULL, code =
   new <- style_token(changes$escaped, changes$href, changes$class)
   out <- replace_in_place(parsed$text, start, end, replacement = new)
 
+  # Add per-line span to match pandoc
+  out <- strsplit(out, "\n")[[1]]
+  out <- paste0("<span>", out, "</span>")
+  out <- paste0(out, collapse = "\n")
+
   if (!is.null(pre_class)) {
     out <- paste0(
       "<pre class='", paste0(pre_class, collapse = " "), "'>\n",
@@ -90,9 +95,21 @@ highlight <- function(text, classes = classes_chroma(), pre_class = NULL, code =
 }
 
 style_token <- function(x, href = NA, class = NA) {
-  x <- ifelse(is.na(href), x, paste0("<a href='", href, "'>", x, "</a>"))
-  x <- ifelse(is.na(class), x, paste0("<span class='", class, "'>", x, "</span>"))
-  x
+  # Split tokens in to lines
+  lines <- strsplit(x, "\n")
+  n <- lengths(lines)
+
+  xs <- unlist(lines)
+  href <- rep(href, n)
+  class <- rep(class, n)
+
+  # Add links and class
+  xs <- ifelse(is.na(href), xs, paste0("<a href='", href, "'>", xs, "</a>"))
+  xs <- ifelse(is.na(class), xs, paste0("<span class='", class, "'>", xs, "</span>"))
+
+  # Re-combine back into lines
+  new_lines <- split(xs, rep(seq_along(x), n))
+  map_chr(new_lines, paste0, collapse = "\n")
 }
 
 # From prettycode:::replace_in_place